@@ -4,25 +4,26 @@ description: A new Flutter project.
 version: 3.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.3 <4.0.0'
 
 dependencies:
-  animations: ^2.0.7
-  cupertino_icons: ^1.0.5
+  animations: ^2.0.11
+  cupertino_icons: ^1.0.6
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.5.0
-  intl: ^0.18.1
-  provider: ^6.0.5
-  shared_preferences: ^2.2.0
+  font_awesome_flutter: ^10.7.0
+  intl: ^0.19.0
+  provider: ^6.1.1
+  shared_preferences: ^2.2.2
 
 dev_dependencies:
   flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^2.0.2
-  flutter_native_splash: ^2.3.1
+  flutter_lints: ^3.0.1
+  flutter_native_splash: ^2.3.8
   flutter_test:
     sdk: flutter
-flutter_icons:
+    
+flutter_launcher_icons:
   android: true
   ios: true
   image_path: "android/app/src/main/res/drawable/launcher_icon.png"
