@@ -12,8 +12,7 @@ class FrostyBackground extends StatelessWidget {
   final Widget child;
 
   const FrostyBackground(
-      {Key? key, this.color, this.intensity = 25, required this.child})
-      : super(key: key);
+      {super.key, this.color, this.intensity = 25, required this.child});
   @override
   Widget build(BuildContext context) {
     return ClipRect(
@@ -44,15 +43,14 @@ class PressableCard extends StatefulWidget {
   final Duration duration;
 
   const PressableCard(
-      {Key? key,
+      {super.key,
       required this.onPressed,
       this.borderRadius = const BorderRadius.all(Radius.circular(5)),
       this.upElevation = 2,
       this.downElevation = 0,
       this.shadowColor = Colors.black,
       this.duration = const Duration(milliseconds: 100),
-      required this.child})
-      : super(key: key);
+      required this.child});
   @override
   _PressableCardState createState() => _PressableCardState();
 }
@@ -101,11 +99,10 @@ class VeggieCard extends StatelessWidget {
   final bool? isPreferredCategory;
 
   const VeggieCard(
-      {Key? key,
+      {super.key,
       required this.veggie,
       required this.isInSeason,
-      this.isPreferredCategory})
-      : super(key: key);
+      this.isPreferredCategory});
 
   Widget _buildDetails() {
     return FrostyBackground(
